+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fix download link in WooCommerce 6.5
+
 ### 1.3.7: November 16th, 2021
 * Duplicate Post compatibility: don't copy _voucheruses when copying product